Remove term name from scraper
The legislature does not have distinct terms.

"Elections in Alderney are held for the positions both of President and Member of the States of Alderney. The President of the States of Alderney is directly elected every four years. Half of the ten States Members are elected every two years for a four-year mandate." [https://en.wikipedia.org/wiki/Elections_in_Alderney]

As there are no set terms, the scraper does not need to record a term name.

(For EP purposes, the term can be set in the import query.)